+++ c_str_span/c_str_result_internal.h
@@ -45,7 +45,7 @@ extern "C" {
 #define _az_RETURN_IF_NOT_ENOUGH_SIZE(span, required_size) \
   do                                                       \
   {                                                        \
-    int32_t const _az_req_sz = (required_size);            \
+    size_t const _az_req_sz = (required_size);             \
     if (az_span_size(span) < _az_req_sz || _az_req_sz < 0) \
     {                                                      \
       return AZ_ERROR_NOT_ENOUGH_SPACE;                    \
